refactor: ObservableModel → EventedModel + 回调自清理替代 destroyed 信号
- ObservableModel → EventedModel，_observers → _callbacks
- 移除 widget.destroyed 连接，避免 PySide6 退出时解释器解体导致 AttributeError
- 改用回调内 weakref 自清理：widget 没了 → 下次触发时 unsubscribe 自身
