+++ src/danmaku_sender/config/monitor_config.py
@@ -1,9 +1,9 @@
 from pydantic import ConfigDict, Field
 
-from danmaku_sender.types.models.observable import ObservableModel
+from danmaku_sender.types.models.evented_model import EventedModel
 
 
-class MonitorConfig(ObservableModel):
+class MonitorConfig(EventedModel):
     """监视器的配置数据"""
     model_config = ConfigDict(validate_assignment=True)
